Elevator Auto
Added two auto commands. One moves the elevator and ejector motors on a two second timer. The other is a parallel command which does the elevator auto and WaitThenForward at the same time. I also added a failsafe to the DriveBaseWait command so it doesn't run when the elevator is moving.